requirements.txt
Django==5.2.4
face_recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
dlib==19.24.2
